moved function MovepieceVisualTeleport into MoveValidatort
The function MovePieceVisualTeleport visual moves a piece on the board. This needs to be called before the actual move to work correctly. Because moveValidator is responsible for moving pieces and checking if its good, it makes sense to put it directly inside that rather than a seperate call before each time movePiece is called. also fixed the game not switching turns after the ai moves